requirements.txt
fastapi
uvicorn
starlette
strawberry-graphql
pydantic
duckdb
nest-asyncio
numpy
pandas